+++ .github/actions/rust-compute-affected/action.yml
@@ -1,0 +1,103 @@
+name: Compute affected Rust services
+description: |
+    Builds the affected-services binary (cached) and runs it to determine
+    which crates and container images are affected by the current change.
+
+inputs:
+    event-name:
+        description: github.event_name — controls how the base ref is resolved
+        required: true
+    pr-base-sha:
+        description: 'Base SHA for PR / merge-group events'
+        required: false
+        default: ''
+    push-before-sha:
+        description: 'Previous HEAD for push events (github.event.before)'
+        required: false
+        default: ''
+    rebuild-all:
+        description: 'Force a full rebuild (skip diff analysis)'
+        required: false
+        default: 'false'
+
+outputs:
+    crates:
+        description: JSON array of affected crate names
+        value: ${{ steps.compute.outputs.crates }}
+    images:
+        description: "Comma-separated affected image names (empty = all, 'none' = skip)"
+        value: ${{ steps.compute.outputs.images }}
+    rebuild-all:
+        description: "'true' when every service needs rebuilding"
+        value: ${{ steps.compute.outputs.rebuild_all }}
+
+runs:
+    using: composite
+    steps:
+        - name: Restore cached binary
+          id: cache
+          uses: actions/cache/restore@9255dc7a253b0ccc959486e2bca901246202afeb # v5.0.1
+          with:
+              path: rust/target/release/affected-services
+              key: affected-services-${{ runner.os }}-${{ runner.arch }}-${{ hashFiles('rust/affected-services/**', 'rust/Cargo.lock') }}
+
+        - name: Build on cache miss
+          if: steps.cache.outputs.cache-hit != 'true'
+          shell: bash
+          working-directory: rust
+          run: cargo build -p affected-services --release --locked
+
+        - name: Save binary to cache
+          if: steps.cache.outputs.cache-hit != 'true'
+          uses: actions/cache/save@9255dc7a253b0ccc959486e2bca901246202afeb # v5.0.1
+          with:
+              path: rust/target/release/affected-services
+              key: affected-services-${{ runner.os }}-${{ runner.arch }}-${{ hashFiles('rust/affected-services/**', 'rust/Cargo.lock') }}
+
+        - name: Compute affected services
+          id: compute
+          shell: bash
+          env:
+              EVENT_NAME: ${{ inputs.event-name }}
+              PR_BASE_SHA: ${{ inputs.pr-base-sha }}
+              PUSH_BEFORE_SHA: ${{ inputs.push-before-sha }}
+              FORCE_REBUILD: ${{ inputs.rebuild-all }}
+          working-directory: rust
+          run: |
+              if [ "$FORCE_REBUILD" = "true" ]; then
+                  echo "rebuild-all requested"
+                  result=$(./target/release/affected-services --files rust/Cargo.lock)
+              else
+                  if [ "$EVENT_NAME" = "push" ]; then
+                      base_ref="$PUSH_BEFORE_SHA"
+                  else
+                      base_ref="$PR_BASE_SHA"
+                  fi
+
+                  if [ -z "$base_ref" ] || echo "$base_ref" | grep -qE '^0+$'; then
+                      echo "No valid base ref — rebuilding all"
+                      result=$(./target/release/affected-services --files rust/Cargo.lock)
+                  else
+                      result=$(./target/release/affected-services --base-ref "$base_ref")
+                  fi
+              fi
+
+              rebuild_all=$(echo "$result" | jq -r '.rebuild_all')
+              crates=$(echo "$result" | jq -c '.crates')
+              images=$(echo "$result" | jq -r '.images | join(",")')
+
+              echo "rebuild_all=$rebuild_all, crates=$crates, images=$images"
+
+              echo "rebuild_all=$rebuild_all" >> "$GITHUB_OUTPUT"
+              echo "crates=$crates" >> "$GITHUB_OUTPUT"
+
+              if [ "$rebuild_all" = "true" ]; then
+                  echo "Rebuild all triggered"
+                  echo "images=" >> "$GITHUB_OUTPUT"
+              elif [ -z "$images" ]; then
+                  echo "No images affected — build will be skipped"
+                  echo "images=none" >> "$GITHUB_OUTPUT"
+              else
+                  echo "Affected images: $images"
+                  echo "images=$images" >> "$GITHUB_OUTPUT"
+              fi
