+++ .github/workflows/rust-docker-build.yml
@@ -52,8 +52,7 @@ jobs:
                       rust/
                       proto/
                       .github/rust-images.yml
-                      .github/actions/setup-protoc/
-                      .github/actions/rust-build-affected-services/
+                      .github/actions/rust-compute-affected/
                   sparse-checkout-cone-mode: false
                   clean: false
 
@@ -62,52 +61,22 @@ jobs:
               with:
                   toolchain: 1.91.1
 
-            - name: Install protoc
-              uses: ./.github/actions/setup-protoc
-
-            - name: Build affected-services tool
-              uses: ./.github/actions/rust-build-affected-services
-
             - name: Compute affected images
               id: compute
-              env:
-                  EVENT_NAME: ${{ github.event_name }}
-                  PR_BASE_SHA: ${{ github.event.pull_request.base.sha }}
-                  PUSH_BEFORE_SHA: ${{ github.event.before }}
-              working-directory: rust
-              run: |
-                  if [ "$EVENT_NAME" = "pull_request" ]; then
-                      base_ref="$PR_BASE_SHA"
-                  else
-                      base_ref="$PUSH_BEFORE_SHA"
-                  fi
-
-                  if [ -z "$base_ref" ] || echo "$base_ref" | grep -qE '^0+$'; then
-                      echo "No valid base ref — rebuilding all"
-                      result=$(./target/release/affected-services --files rust/Cargo.lock)
-                  else
-                      result=$(./target/release/affected-services --base-ref "$base_ref")
-                  fi
-                  rebuild_all=$(echo "$result" | jq -r '.rebuild_all')
-                  images=$(echo "$result" | jq -r '.images | join(",")')
-
-                  if [ "$rebuild_all" = "true" ]; then
-                      echo "Rebuild all triggered"
-                      echo "images=" >> "$GITHUB_OUTPUT"
-                  elif [ -z "$images" ]; then
-                      echo "No images affected — build will be skipped"
-                      echo "images=none" >> "$GITHUB_OUTPUT"
-                  else
-                      echo "Affected images: $images"
-                      echo "images=$images" >> "$GITHUB_OUTPUT"
-                  fi
+              uses: ./.github/actions/rust-compute-affected
+              with:
+                  event-name: ${{ github.event_name }}
+                  pr-base-sha: ${{ github.event.pull_request.base.sha }}
+                  push-before-sha: ${{ github.event.before }}
 
     build-images:
         needs: affected
         # Run unless affected explicitly found zero images.
-        # When affected is skipped (workflow_dispatch), filter is empty = build all.
+        # When affected is skipped (workflow_dispatch), result is 'skipped' = build all.
+        # When affected fails (tool panic, checkout error), result is 'failure' = stop.
         if: |
             always() && !cancelled() &&
+            needs.affected.result != 'failure' &&
             needs.affected.outputs.image-filter != 'none'
         uses: ./.github/workflows/_rust-build-images.yml
         with:
