+++ rust/affected-services/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod affected;
+pub mod dump;
+pub mod graph;
+pub mod images;
