+++ rust/affected-services/src/main.rs
@@ -1,46 +1,15 @@
-use std::collections::{HashMap, HashSet};
 use std::io::BufRead;
-use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::path::Path;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use clap::{Parser, ValueEnum};
-use determinator::rules::DeterminatorRules;
-use determinator::Determinator;
-use guppy::graph::{DependencyDirection, PackageGraph};
-use guppy::MetadataCommand;
-use serde::Serialize;
 
-const RULES_TOML: &str = include_str!("../determinator-rules.toml");
-
-const WORKSPACE_PREFIX: &str = "rust/";
-
-const NON_CRATE_IMAGE_TRIGGERS: &[(&str, &[&str])] = &[(
-    "sqlx-migrate",
-    &[
-        "rust/persons_migrations/",
-        "rust/behavioral_cohorts_migrations/",
-        "rust/cyclotron-core/migrations/",
-        "rust/cyclotron-node-migrations/",
-        "rust/flags_read_store_migrations/",
-        "rust/bin/migrate-",
-        "rust/Dockerfile.sqlx-migrate",
-    ],
-)];
-
-#[derive(Debug, Default)]
-struct ImageConfig {
-    image: String,
-    bin: Option<String>,
-}
-
-#[derive(Debug, Serialize)]
-struct AffectedResult {
-    rebuild_all: bool,
-    images: Vec<String>,
-    crates: Vec<String>,
-    directly_changed: Vec<String>,
-}
+use affected_services::affected::compute_affected;
+use affected_services::dump::dump_graph;
+use affected_services::graph::{
+    build_old_package_graph, build_package_graph, find_repo_root, get_changed_files,
+};
+use affected_services::images::parse_images_yaml;
 
 #[derive(Parser)]
 #[command(about = "Compute affected Rust services from changed files")]
@@ -58,10 +27,10 @@ struct Cli {
     dump_graph: bool,
 
     #[arg(long)]
-    workspace_dir: Option<PathBuf>,
+    workspace_dir: Option<std::path::PathBuf>,
 
     #[arg(long)]
-    images_file: Option<PathBuf>,
+    images_file: Option<std::path::PathBuf>,
 
     #[arg(long, value_enum, default_value = "json")]
     output: OutputFormat,
@@ -73,427 +42,6 @@ enum OutputFormat {
     Images,
     Crates,
 }
-
-// ── I/O ────────────────────────────────────────────────────────────
-
-fn find_repo_root() -> Result<PathBuf> {
-    let out = Command::new("git")
-        .args(["rev-parse", "--show-toplevel"])
-        .output()
-        .context("failed to run git")?;
-    Ok(PathBuf::from(
-        String::from_utf8(out.stdout)?.trim().to_string(),
-    ))
-}
-
-fn get_changed_files(base_ref: &str) -> Result<Vec<String>> {
-    let out = Command::new("git")
-        .args(["diff", "--name-only", &format!("{base_ref}...HEAD")])
-        .output()
-        .context("failed to run git diff")?;
-    anyhow::ensure!(
-        out.status.success(),
-        "git diff exited with {}: {}",
-        out.status,
-        String::from_utf8_lossy(&out.stderr).trim()
-    );
-    Ok(String::from_utf8(out.stdout)?
-        .lines()
-        .filter(|l| !l.is_empty())
-        .map(String::from)
-        .collect())
-}
-
-fn build_package_graph(workspace_dir: &Path) -> Result<PackageGraph> {
-    let mut cmd = MetadataCommand::new();
-    cmd.current_dir(workspace_dir);
-    cmd.build_graph()
-        .context("failed to build package graph from cargo metadata")
-}
-
-// ── Old graph from base revision ──────────────────────────────────
-
-struct TempWorktree {
-    path: PathBuf,
-}
-
-impl TempWorktree {
-    fn create(base_ref: &str) -> Result<Self> {
-        let path = tempfile::Builder::new()
-            .prefix("affected-services-")
-            .tempdir()
-            .context("failed to create temp directory")?
-            .keep();
-        let out = Command::new("git")
-            .args([
-                "worktree",
-                "add",
-                "--detach",
-                &path.to_string_lossy(),
-                base_ref,
-            ])
-            .output()
-            .context("failed to create git worktree")?;
-        anyhow::ensure!(
-            out.status.success(),
-            "git worktree add failed: {}",
-            String::from_utf8_lossy(&out.stderr).trim()
-        );
-        Ok(Self { path })
-    }
-
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for TempWorktree {
-    fn drop(&mut self) {
-        let _ = Command::new("git")
-            .args([
-                "worktree",
-                "remove",
-                "--force",
-                &self.path.to_string_lossy(),
-            ])
-            .status();
-    }
-}
-
-fn build_old_package_graph(base_ref: &str, workspace_subdir: &str) -> Option<PackageGraph> {
-    let worktree = match TempWorktree::create(base_ref) {
-        Ok(w) => w,
-        Err(e) => {
-            eprintln!("warning: could not create worktree at {base_ref}: {e}");
-            return None;
-        }
-    };
-
-    let old_workspace = worktree.path().join(workspace_subdir);
-    if !old_workspace.join("Cargo.toml").exists() {
-        eprintln!("warning: {base_ref} has no {workspace_subdir}/Cargo.toml");
-        return None;
-    }
-
-    let mut cmd = MetadataCommand::new();
-    cmd.current_dir(&old_workspace);
-    cmd.other_options(["--locked"]);
-    match cmd.build_graph() {
-        Ok(graph) => Some(graph),
-        Err(e) => {
-            eprintln!("warning: could not build old package graph: {e}");
-            None
-        }
-    }
-}
-
-fn parse_images_yaml(path: &Path) -> Result<Vec<ImageConfig>> {
-    let content = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read {}", path.display()))?;
-    parse_images_yaml_content(&content)
-}
-
-fn parse_images_yaml_content(content: &str) -> Result<Vec<ImageConfig>> {
-    let mut items = Vec::new();
-    let mut current: Option<ImageConfig> = None;
-
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        if line.starts_with("- ") {
-            if let Some(item) = current.take() {
-                items.push(item);
-            }
-            current = Some(ImageConfig::default());
-            let rest = trimmed.strip_prefix("- ").unwrap_or(trimmed);
-            if let Some(item) = current.as_mut() {
-                apply_field(item, rest);
-            }
-        } else if let Some(ref mut item) = current {
-            apply_field(item, trimmed);
-        }
-    }
-    if let Some(item) = current {
-        items.push(item);
-    }
-    Ok(items)
-}
-
-fn apply_field(config: &mut ImageConfig, line: &str) {
-    if let Some((key, value)) = line.split_once(':') {
-        let value = value.trim();
-        match key.trim() {
-            "image" => config.image = value.to_string(),
-            "bin" => config.bin = Some(value.to_string()),
-            _ => {}
-        }
-    }
-}
-
-// ── Path transformation ───────────────────────────────────────────
-
-/// Convert repo-root-relative paths to workspace-relative paths.
-///
-/// `rust/capture/src/main.rs` → `capture/src/main.rs`
-/// `proto/personhog.proto`    → `../proto/personhog.proto`
-fn to_workspace_relative(repo_path: &str) -> String {
-    if let Some(rest) = repo_path.strip_prefix(WORKSPACE_PREFIX) {
-        rest.to_string()
-    } else {
-        format!("../{repo_path}")
-    }
-}
-
-// ── Image mapping ─────────────────────────────────────────────────
-
-fn build_binary_to_crate_map(graph: &PackageGraph) -> HashMap<String, String> {
-    let mut map = HashMap::new();
-    for member in graph.workspace().iter() {
-        for target in member.build_targets() {
-            if let guppy::graph::BuildTargetId::Binary(name) = target.id() {
-                map.insert(name.to_string(), member.name().to_string());
-            }
-        }
-    }
-    map
-}
-
-fn map_crates_to_images(
-    affected: &HashSet<String>,
-    bin_to_crate: &HashMap<String, String>,
-    images: &[ImageConfig],
-) -> Vec<String> {
-    let mut result: Vec<String> = images
-        .iter()
-        .filter_map(|img| {
-            let bin_name = img.bin.as_deref().unwrap_or(&img.image);
-            let owning_crate = bin_to_crate.get(bin_name)?;
-            affected.contains(owning_crate).then(|| img.image.clone())
-        })
-        .collect();
-    result.sort();
-    result
-}
-
-fn non_crate_images_affected(files: &[String]) -> Vec<String> {
-    let mut result = Vec::new();
-    for &(image, prefixes) in NON_CRATE_IMAGE_TRIGGERS {
-        if files
-            .iter()
-            .any(|f| prefixes.iter().any(|p| f.starts_with(p)))
-        {
-            result.push(image.to_string());
-        }
-    }
-    result
-}
-
-// ── Core logic ────────────────────────────────────────────────────
-
-fn compute_affected(
-    changed_files: &[String],
-    old_graph: Option<&PackageGraph>,
-    new_graph: &PackageGraph,
-    images: &[ImageConfig],
-) -> Result<AffectedResult> {
-    let rules =
-        DeterminatorRules::parse(RULES_TOML).context("failed to parse determinator rules")?;
-
-    let workspace_paths: Vec<String> = changed_files
-        .iter()
-        .map(|f| to_workspace_relative(f))
-        .collect();
-
-    let old = old_graph.unwrap_or(new_graph);
-    let mut det = Determinator::new(old, new_graph);
-    det.set_rules(&rules)
-        .context("failed to set determinator rules")?;
-    det.add_changed_paths(workspace_paths.iter().map(|s| s.as_str()));
-    let result = det.compute();
-
-    let workspace_size = new_graph.workspace().iter().count();
-    let affected_count = result.affected_set.len();
-    let rebuild_all = affected_count == workspace_size;
-
-    let bin_to_crate = build_binary_to_crate_map(new_graph);
-
-    let affected_names: HashSet<String> = result
-        .affected_set
-        .packages(DependencyDirection::Forward)
-        .map(|p| p.name().to_string())
-        .collect();
-
-    let directly_changed_names: HashSet<String> = result
-        .path_changed_set
-        .packages(DependencyDirection::Forward)
-        .map(|p| p.name().to_string())
-        .collect();
-
-    if rebuild_all {
-        let mut all_images: Vec<String> = images.iter().map(|i| i.image.clone()).collect();
-        all_images.sort();
-        let mut all_crates: Vec<String> = affected_names.into_iter().collect();
-        all_crates.sort();
-        return Ok(AffectedResult {
-            rebuild_all: true,
-            images: all_images,
-            crates: all_crates,
-            directly_changed: Vec::new(),
-        });
-    }
-
-    let mut affected_images = map_crates_to_images(&affected_names, &bin_to_crate, images);
-    for img in non_crate_images_affected(changed_files) {
-        if !affected_images.contains(&img) {
-            affected_images.push(img);
-        }
-    }
-    affected_images.sort();
-
-    let mut crates: Vec<String> = affected_names.into_iter().collect();
-    crates.sort();
-    let mut directly: Vec<String> = directly_changed_names.into_iter().collect();
-    directly.sort();
-
-    Ok(AffectedResult {
-        rebuild_all: false,
-        images: affected_images,
-        crates,
-        directly_changed: directly,
-    })
-}
-
-// ── Dump mode ─────────────────────────────────────────────────────
-
-fn dump_graph(graph: &PackageGraph, images: &[ImageConfig]) {
-    let bin_to_crate = build_binary_to_crate_map(graph);
-
-    let mut crate_to_bins: HashMap<&str, Vec<&str>> = HashMap::new();
-    for (bin_name, crate_name) in &bin_to_crate {
-        crate_to_bins
-            .entry(crate_name.as_str())
-            .or_default()
-            .push(bin_name.as_str());
-    }
-
-    let deployable_crates: HashSet<&str> = images
-        .iter()
-        .filter_map(|img| {
-            let bin_name = img.bin.as_deref().unwrap_or(&img.image);
-            bin_to_crate.get(bin_name).map(|s| s.as_str())
-        })
-        .collect();
-
-    println!("REVERSE DEPENDENCY GRAPH");
-    println!("(crate → workspace crates that depend on it)");
-    println!("{}", "=".repeat(70));
-
-    let mut members: Vec<_> = graph.workspace().iter().collect();
-    members.sort_by_key(|m| m.name().to_string());
-
-    for member in &members {
-        let name = member.name();
-
-        let mut dependents: Vec<String> = Vec::new();
-        for link in member.reverse_direct_links() {
-            let from = link.from();
-            if !graph.workspace().contains_name(from.name()) {
-                continue;
-            }
-            let is_dev_only = !link.normal().is_present() && !link.build().is_present();
-            if !is_dev_only {
-                dependents.push(from.name().to_string());
-            }
-        }
-        dependents.sort();
-
-        let mut tags = Vec::new();
-        if let Some(bins) = crate_to_bins.get(name) {
-            let mut bins = bins.clone();
-            bins.sort();
-            tags.push(format!("bins: {}", bins.join(", ")));
-        }
-        if deployable_crates.contains(name) {
-            tags.push("deployable".to_string());
-        }
-        let tag_str = if tags.is_empty() {
-            String::new()
-        } else {
-            format!("  [{}]", tags.join(", "))
-        };
-
-        let dep_str = if dependents.is_empty() {
-            "(leaf)".to_string()
-        } else {
-            dependents.join(", ")
-        };
-
-        println!("  {name}{tag_str}");
-        println!("    depended on by: {dep_str}");
-    }
-
-    println!();
-    println!("TRANSITIVE IMPACT (crates affecting >2 others)");
-    println!("{}", "=".repeat(70));
-
-    let rules = DeterminatorRules::parse(RULES_TOML).expect("rules should parse");
-
-    let mut impacts: Vec<(&str, usize, Vec<String>)> = Vec::new();
-    for member in &members {
-        let name = member.name();
-        let workspace_path = format!("{}/src/lib.rs", member.source().workspace_path().unwrap());
-        let mut det = Determinator::new(graph, graph);
-        det.set_rules(&rules).unwrap();
-        det.add_changed_paths(std::iter::once(workspace_path.as_str()));
-        let result = det.compute();
-
-        let affected: HashSet<String> = result
-            .affected_set
-            .packages(DependencyDirection::Forward)
-            .map(|p| p.name().to_string())
-            .collect();
-        let affected_images = map_crates_to_images(&affected, &bin_to_crate, images);
-        let dep_count = affected.len().saturating_sub(1);
-        if dep_count > 2 {
-            impacts.push((name, dep_count, affected_images));
-        }
-    }
-    impacts.sort_by(|a, b| b.1.cmp(&a.1));
-
-    for (crate_name, dep_count, imgs) in &impacts {
-        println!("  {crate_name} → {dep_count} crates, {} images", imgs.len());
-        if !imgs.is_empty() {
-            println!("    images: {}", imgs.join(", "));
-        }
-    }
-
-    println!();
-    println!("BINARY → CRATE → IMAGE MAPPING");
-    println!("{}", "=".repeat(70));
-
-    let mut sorted_images: Vec<&ImageConfig> = images.iter().collect();
-    sorted_images.sort_by_key(|i| &i.image);
-
-    for img in sorted_images {
-        let bin_name = img.bin.as_deref().unwrap_or(&img.image);
-        let owning = bin_to_crate
-            .get(bin_name)
-            .map(|s| s.as_str())
-            .unwrap_or("???");
-        let mut extra = String::new();
-        if img.bin.is_some() {
-            extra.push_str(&format!("  (bin={bin_name})"));
-        }
-        println!(
-            "  image:{}  ←  crate:{owning}  ←  bin:{bin_name}{extra}",
-            img.image
-        );
-    }
-}
-
-// ── Main ──────────────────────────────────────────────────────────
 
 fn main() -> Result<()> {
     let cli = Cli::parse();
@@ -543,531 +91,4 @@ fn main() -> Result<()> {
         OutputFormat::Crates => println!("{}", serde_json::to_string(&result.crates)?),
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn workspace_relative_strips_rust_prefix() {
-        assert_eq!(
-            to_workspace_relative("rust/capture/src/main.rs"),
-            "capture/src/main.rs"
-        );
-    }
-
-    #[test]
-    fn workspace_relative_adds_parent_for_external() {
-        assert_eq!(
-            to_workspace_relative("proto/personhog.proto"),
-            "../proto/personhog.proto"
-        );
-    }
-
-    #[test]
-    fn workspace_relative_github_files() {
-        assert_eq!(
-            to_workspace_relative(".github/rust-images.yml"),
-            "../.github/rust-images.yml"
-        );
-    }
-
-    #[test]
-    fn non_crate_migration_trigger() {
-        let affected = non_crate_images_affected(&["rust/persons_migrations/new.sql".into()]);
-        assert_eq!(affected, vec!["sqlx-migrate"]);
-    }
-
-    #[test]
-    fn non_crate_no_false_positives() {
-        let affected = non_crate_images_affected(&["rust/capture/src/main.rs".into()]);
-        assert!(affected.is_empty());
-    }
-
-    #[test]
-    fn image_mapping() {
-        let affected = HashSet::from(["feature-flags".into()]);
-        let bin_to_crate = HashMap::from([
-            ("feature-flags".into(), "feature-flags".into()),
-            ("warm-flags-cache".into(), "feature-flags".into()),
-            ("capture".into(), "capture".into()),
-        ]);
-        let images = vec![
-            ImageConfig {
-                image: "feature-flags".into(),
-                bin: None,
-            },
-            ImageConfig {
-                image: "flags-cache-warmer".into(),
-                bin: Some("warm-flags-cache".into()),
-            },
-            ImageConfig {
-                image: "capture".into(),
-                bin: None,
-            },
-        ];
-
-        let result = map_crates_to_images(&affected, &bin_to_crate, &images);
-        assert_eq!(result, vec!["feature-flags", "flags-cache-warmer"]);
-    }
-
-    #[test]
-    fn rules_parse_successfully() {
-        DeterminatorRules::parse(RULES_TOML).expect("rules TOML should parse");
-    }
-
-    // ── parse_images_yaml_content unit tests ─────────────────────
-
-    #[test]
-    fn parse_images_simple_entry() {
-        let yaml = "- image: capture\n  dockerfile: ./rust/Dockerfile\n";
-        let items = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].image, "capture");
-        assert!(items[0].bin.is_none());
-    }
-
-    #[test]
-    fn parse_images_with_bin_override() {
-        let yaml = "- image: flags-cache-warmer\n  bin: warm-flags-cache\n";
-        let items = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].image, "flags-cache-warmer");
-        assert_eq!(items[0].bin.as_deref(), Some("warm-flags-cache"));
-    }
-
-    #[test]
-    fn parse_images_multiple_entries() {
-        let yaml = "\
-- image: capture
-  dockerfile: ./rust/Dockerfile
-
-- image: cymbal
-  dockerfile: ./rust/Dockerfile
-
-- image: flags-cache-warmer
-  bin: warm-flags-cache
-  dockerfile: ./rust/Dockerfile
-";
-        let items = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(items.len(), 3);
-        assert_eq!(items[0].image, "capture");
-        assert_eq!(items[1].image, "cymbal");
-        assert_eq!(items[2].image, "flags-cache-warmer");
-        assert_eq!(items[2].bin.as_deref(), Some("warm-flags-cache"));
-    }
-
-    #[test]
-    fn parse_images_skips_comments_and_blank_lines() {
-        let yaml = "\
-# This is a comment
-- image: capture
-
-  # inline comment
-  dockerfile: ./rust/Dockerfile
-
-";
-        let items = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].image, "capture");
-    }
-
-    #[test]
-    fn parse_images_ignores_unknown_fields() {
-        let yaml = "- image: capture\n  dockerfile: ./rust/Dockerfile\n  project: abc123\n  features: some-feature\n  target: runtime-jumphost\n";
-        let items = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].image, "capture");
-        assert!(items[0].bin.is_none());
-    }
-
-    #[test]
-    fn parse_images_empty_input() {
-        let items = parse_images_yaml_content("").unwrap();
-        assert!(items.is_empty());
-    }
-
-    #[test]
-    fn parse_images_comments_only() {
-        let items = parse_images_yaml_content("# just a comment\n# another one\n").unwrap();
-        assert!(items.is_empty());
-    }
-
-    // ── Integration tests against real workspace ──────────────────
-    //
-    // Two kinds of integration tests:
-    //
-    // 1. Structural — assert pipeline invariants that hold regardless of
-    //    which services exist (subset relationships, union property, etc.)
-    //
-    // 2. Concrete — assert against the actual rust-images.yml and workspace.
-    //    These intentionally break when a service is added/renamed/removed
-    //    so the author verifies the wiring is correct.
-
-    fn load_test_fixtures() -> (PackageGraph, Vec<ImageConfig>) {
-        let repo_root = find_repo_root().expect("must be in a git repo");
-        let graph = build_package_graph(&repo_root.join("rust")).expect("cargo metadata");
-        let images =
-            parse_images_yaml(&repo_root.join(".github/rust-images.yml")).expect("images YAML");
-        (graph, images)
-    }
-
-    // ── Structural tests ─────────────────────────────────────────
-
-    #[test]
-    fn e2e_file_in_workspace_crate_affects_that_crate() {
-        let (graph, images) = load_test_fixtures();
-        let member = graph.workspace().iter().next().unwrap();
-        let fake_path = format!(
-            "rust/{}/src/lib.rs",
-            member.source().workspace_path().unwrap()
-        );
-        let result = compute_affected(&[fake_path], None, &graph, &images).unwrap();
-        assert!(!result.rebuild_all);
-        assert!(
-            result.crates.contains(&member.name().to_string()),
-            "changing a file in {} should mark it affected",
-            member.name()
-        );
-    }
-
-    #[test]
-    fn e2e_directly_changed_is_subset_of_crates() {
-        let (graph, images) = load_test_fixtures();
-        let member = graph.workspace().iter().next().unwrap();
-        let fake_path = format!(
-            "rust/{}/src/lib.rs",
-            member.source().workspace_path().unwrap()
-        );
-        let result = compute_affected(&[fake_path], None, &graph, &images).unwrap();
-        let crates: HashSet<&str> = result.crates.iter().map(|s| s.as_str()).collect();
-        for dc in &result.directly_changed {
-            assert!(
-                crates.contains(dc.as_str()),
-                "directly_changed entry {dc} must appear in crates"
-            );
-        }
-    }
-
-    #[test]
-    fn e2e_rebuild_all_marks_every_workspace_crate() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(&["rust/Cargo.lock".into()], None, &graph, &images).unwrap();
-        assert!(result.rebuild_all);
-        let workspace_count = graph.workspace().iter().count();
-        assert_eq!(
-            result.crates.len(),
-            workspace_count,
-            "rebuild_all should include all {} workspace crates",
-            workspace_count
-        );
-    }
-
-    #[test]
-    fn e2e_file_outside_workspace_produces_no_crates() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(&["README.md".into()], None, &graph, &images).unwrap();
-        assert!(!result.rebuild_all);
-        assert!(result.crates.is_empty());
-        assert!(result.images.is_empty());
-    }
-
-    #[test]
-    fn e2e_multi_file_is_union() {
-        let (graph, images) = load_test_fixtures();
-        let members: Vec<_> = graph.workspace().iter().collect();
-        let (a, b) = (&members[0], &members[1]);
-        let path_a = format!("rust/{}/src/lib.rs", a.source().workspace_path().unwrap());
-        let path_b = format!("rust/{}/src/lib.rs", b.source().workspace_path().unwrap());
-
-        let result_a =
-            compute_affected(std::slice::from_ref(&path_a), None, &graph, &images).unwrap();
-        let result_b =
-            compute_affected(std::slice::from_ref(&path_b), None, &graph, &images).unwrap();
-        let result_both = compute_affected(&[path_a, path_b], None, &graph, &images).unwrap();
-
-        let union: HashSet<&str> = result_a
-            .crates
-            .iter()
-            .chain(&result_b.crates)
-            .map(|s| s.as_str())
-            .collect();
-        let combined: HashSet<&str> = result_both.crates.iter().map(|s| s.as_str()).collect();
-        assert_eq!(
-            union, combined,
-            "multi-file result should be the union of individual results"
-        );
-    }
-
-    // ── Concrete tests ───────────────────────────────────────────
-    //
-    // Update these when adding/removing/renaming services.
-
-    /// Every image in rust-images.yml must resolve to a binary target in
-    /// the workspace (except sqlx-migrate which is a non-crate image).
-    #[test]
-    fn every_image_resolves_to_a_workspace_binary() {
-        let (graph, images) = load_test_fixtures();
-        let bin_to_crate = build_binary_to_crate_map(&graph);
-        for img in &images {
-            let bin_name = img.bin.as_deref().unwrap_or(&img.image);
-            if NON_CRATE_IMAGE_TRIGGERS
-                .iter()
-                .any(|&(name, _)| name == img.image)
-            {
-                continue;
-            }
-            assert!(
-                bin_to_crate.contains_key(bin_name),
-                "image '{}' references binary '{}' which doesn't exist in the workspace",
-                img.image,
-                bin_name,
-            );
-        }
-    }
-
-    #[test]
-    fn e2e_leaf_service_only_affects_itself_and_dependents() {
-        let (graph, images) = load_test_fixtures();
-        let result =
-            compute_affected(&["rust/capture/src/main.rs".into()], None, &graph, &images).unwrap();
-        assert!(!result.rebuild_all);
-        assert_eq!(result.directly_changed, vec!["capture"]);
-        assert_eq!(result.images, vec!["capture", "capture-logs"]);
-    }
-
-    #[test]
-    fn e2e_proto_change_maps_to_proto_crates() {
-        let (graph, images) = load_test_fixtures();
-        let result =
-            compute_affected(&["proto/personhog.proto".into()], None, &graph, &images).unwrap();
-        assert!(!result.rebuild_all);
-        assert!(result.directly_changed.contains(&"personhog-proto".into()));
-        assert!(result
-            .directly_changed
-            .contains(&"kafka-assigner-proto".into()));
-        assert!(
-            result.images.len() > 2,
-            "proto change should propagate to multiple images"
-        );
-    }
-
-    #[test]
-    fn e2e_migration_only_triggers_sqlx_migrate() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(
-            &["rust/persons_migrations/new.sql".into()],
-            None,
-            &graph,
-            &images,
-        )
-        .unwrap();
-        assert!(!result.rebuild_all);
-        assert_eq!(result.images, vec!["sqlx-migrate"]);
-        assert!(result.crates.is_empty());
-    }
-
-    #[test]
-    fn e2e_multi_binary_crate_produces_all_images() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(
-            &["rust/feature-flags/src/lib.rs".into()],
-            None,
-            &graph,
-            &images,
-        )
-        .unwrap();
-        assert!(!result.rebuild_all);
-        assert!(result.images.contains(&"feature-flags".into()));
-        assert!(result.images.contains(&"flags-cache-warmer".into()));
-    }
-
-    // ── Determinator two-graph tests ─────────────────────────────
-    //
-    // These exercise the `old_graph` path in `compute_affected`,
-    // which all other tests skip by passing `None`.
-
-    #[test]
-    fn e2e_same_old_and_new_graph_matches_none_fallback() {
-        let (graph, images) = load_test_fixtures();
-        let changed = vec!["rust/capture/src/main.rs".into()];
-
-        let result_none = compute_affected(&changed, None, &graph, &images).unwrap();
-        let result_some = compute_affected(&changed, Some(&graph), &graph, &images).unwrap();
-
-        assert_eq!(result_none.crates, result_some.crates);
-        assert_eq!(result_none.images, result_some.images);
-        assert_eq!(result_none.rebuild_all, result_some.rebuild_all);
-        assert_eq!(result_none.directly_changed, result_some.directly_changed);
-    }
-
-    #[test]
-    fn e2e_old_graph_no_changed_files_produces_empty() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(&[], Some(&graph), &graph, &images).unwrap();
-        assert!(!result.rebuild_all);
-        assert!(result.crates.is_empty());
-        assert!(result.images.is_empty());
-    }
-
-    #[test]
-    fn e2e_old_graph_rebuild_all_on_lockfile() {
-        let (graph, images) = load_test_fixtures();
-        let result =
-            compute_affected(&["rust/Cargo.lock".into()], Some(&graph), &graph, &images).unwrap();
-        assert!(result.rebuild_all);
-        let workspace_count = graph.workspace().iter().count();
-        assert_eq!(result.crates.len(), workspace_count);
-    }
-
-    #[test]
-    fn e2e_old_graph_leaf_service() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(
-            &["rust/capture/src/main.rs".into()],
-            Some(&graph),
-            &graph,
-            &images,
-        )
-        .unwrap();
-        assert!(!result.rebuild_all);
-        assert_eq!(result.directly_changed, vec!["capture"]);
-        assert!(result.crates.contains(&"capture".into()));
-    }
-
-    #[test]
-    fn e2e_old_graph_proto_propagation() {
-        let (graph, images) = load_test_fixtures();
-        let result = compute_affected(
-            &["proto/personhog.proto".into()],
-            Some(&graph),
-            &graph,
-            &images,
-        )
-        .unwrap();
-        assert!(!result.rebuild_all);
-        assert!(result.directly_changed.contains(&"personhog-proto".into()));
-        assert!(
-            result.images.len() > 2,
-            "proto change should propagate to multiple images via old_graph path"
-        );
-    }
-
-    #[test]
-    fn e2e_old_graph_from_master() {
-        let repo_root = find_repo_root().expect("must be in a git repo");
-        let old_graph = match build_old_package_graph("master", "rust") {
-            Some(g) => g,
-            None => {
-                eprintln!("skipping: could not build old graph from master");
-                return;
-            }
-        };
-        let new_graph = build_package_graph(&repo_root.join("rust")).expect("cargo metadata");
-        let images =
-            parse_images_yaml(&repo_root.join(".github/rust-images.yml")).expect("images YAML");
-
-        let result = compute_affected(
-            &["rust/capture/src/main.rs".into()],
-            Some(&old_graph),
-            &new_graph,
-            &images,
-        )
-        .unwrap();
-        assert!(result.crates.contains(&"capture".into()));
-    }
-
-    // ── YAML parsing unit tests ─────────────────────────────────
-
-    #[test]
-    fn parse_yaml_single_image() {
-        let yaml = "- image: capture\n  dockerfile: ./rust/Dockerfile\n  project: abc123\n";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 1);
-        assert_eq!(images[0].image, "capture");
-        assert_eq!(images[0].bin, None);
-    }
-
-    #[test]
-    fn parse_yaml_image_with_bin() {
-        let yaml = "- image: flags-cache-warmer\n  bin: warm-flags-cache\n  dockerfile: ./rust/Dockerfile\n";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 1);
-        assert_eq!(images[0].image, "flags-cache-warmer");
-        assert_eq!(images[0].bin.as_deref(), Some("warm-flags-cache"));
-    }
-
-    #[test]
-    fn parse_yaml_multiple_images() {
-        let yaml = "\
-- image: capture
-  dockerfile: ./rust/Dockerfile
-
-- image: cymbal
-  dockerfile: ./rust/Dockerfile
-
-- image: feature-flags
-  bin: feature-flags
-  dockerfile: ./rust/Dockerfile
-";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 3);
-        assert_eq!(images[0].image, "capture");
-        assert_eq!(images[1].image, "cymbal");
-        assert_eq!(images[2].image, "feature-flags");
-    }
-
-    #[test]
-    fn parse_yaml_skips_comments_and_blank_lines() {
-        let yaml = "\
-# This is a comment
-- image: capture
-  dockerfile: ./rust/Dockerfile
-
-# Another comment
-
-- image: cymbal
-  dockerfile: ./rust/Dockerfile
-";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 2);
-        assert_eq!(images[0].image, "capture");
-        assert_eq!(images[1].image, "cymbal");
-    }
-
-    #[test]
-    fn parse_yaml_empty_input() {
-        let images = parse_images_yaml_content("").unwrap();
-        assert!(images.is_empty());
-    }
-
-    #[test]
-    fn parse_yaml_comments_only() {
-        let images = parse_images_yaml_content("# just a comment\n# another\n").unwrap();
-        assert!(images.is_empty());
-    }
-
-    #[test]
-    fn parse_yaml_unknown_fields_ignored() {
-        let yaml = "\
-- image: capture
-  dockerfile: ./rust/Dockerfile
-  project: abc123
-  target: runtime-jumphost
-  features: warm-flags-cache
-";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 1);
-        assert_eq!(images[0].image, "capture");
-        assert_eq!(images[0].bin, None);
-    }
-
-    #[test]
-    fn parse_yaml_inline_image_on_dash_line() {
-        let yaml = "- image: capture\n";
-        let images = parse_images_yaml_content(yaml).unwrap();
-        assert_eq!(images.len(), 1);
-        assert_eq!(images[0].image, "capture");
-    }
 }
